Move and rename invoice purger files
